swaybar: feactor render, statusline
